Updated test stacks
When creating the time-grids using the new (faster) interpolation,
there are minor changes in the times, that create minor changes in the
final migration.

As the new interpolation from scipy.ndimage is much better than the
original, hand-written one, I trust the new time-grids more than the
old ones. So ok for updating the stack signature files.